+++ projects/net.splitcells.dem.merger/src/main/java/net/splitcells/dem/resource/communication/interaction/LogLevelTest.java
@@ -22,8 +22,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class LogLevelTest {
     @Test
     public void testPrioritySmallerThanOrder() {
-        assertThat(LogLevel.ERROR.smallerThan(LogLevel.DEBUG)).isTrue();
+        assertThat(LogLevel.ERROR.smallerThan(LogLevel.DEBUG)).isFalse();
         assertThat(LogLevel.DEBUG.smallerThan(LogLevel.DEBUG)).isFalse();
-        assertThat(LogLevel.DEBUG.smallerThan(LogLevel.ERROR)).isFalse();
+        assertThat(LogLevel.DEBUG.smallerThan(LogLevel.ERROR)).isTrue();
     }
 }